+++ backend/src/main/java/dev/tripdraw/domain/post/Post.java
@@ -6,7 +6,7 @@ import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
 
 import dev.tripdraw.domain.common.BaseEntity;
-import dev.tripdraw.domain.member.Member;
+import dev.tripdraw.member.domain.Member;
 import dev.tripdraw.trip.domain.Point;
 import dev.tripdraw.exception.post.PostException;
 import jakarta.persistence.Column;
