+++ backend/src/main/java/dev/tripdraw/dto/post/PostAndPointCreateRequest.java
@@ -3,7 +3,7 @@ package dev.tripdraw.dto.post;
 import static com.fasterxml.jackson.annotation.JsonFormat.Shape.STRING;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import dev.tripdraw.domain.member.Member;
+import dev.tripdraw.member.domain.Member;
 import dev.tripdraw.domain.post.Post;
 import dev.tripdraw.trip.domain.Point;
 import io.swagger.v3.oas.annotations.media.Schema;
