+++ backend/src/main/java/dev/tripdraw/dto/post/PostRequest.java
@@ -1,6 +1,6 @@
 package dev.tripdraw.dto.post;
 
-import dev.tripdraw.domain.member.Member;
+import dev.tripdraw.member.domain.Member;
 import dev.tripdraw.domain.post.Post;
 import dev.tripdraw.trip.domain.Point;
 import io.swagger.v3.oas.annotations.media.Schema;
