+++ backend/src/main/java/dev/tripdraw/member/domain/MemberRepository.java
@@ -1,9 +1,9 @@
-package dev.tripdraw.domain.member;
+package dev.tripdraw.member.domain;
 
-import static dev.tripdraw.exception.member.MemberExceptionType.MEMBER_NOT_FOUND;
+import static dev.tripdraw.member.exception.MemberExceptionType.MEMBER_NOT_FOUND;
 
 import dev.tripdraw.domain.oauth.OauthType;
-import dev.tripdraw.exception.member.MemberException;
+import dev.tripdraw.member.exception.MemberException;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.lang.NonNull;
