+++ backend/src/main/java/dev/tripdraw/member/exception/MemberException.java
@@ -1,4 +1,4 @@
-package dev.tripdraw.exception.member;
+package dev.tripdraw.member.exception;
 
 import dev.tripdraw.exception.common.BaseException;
 import dev.tripdraw.exception.common.ExceptionType;
