+++ backend/src/main/java/dev/tripdraw/presentation/member/AuthInterceptor.java
@@ -2,7 +2,7 @@ package dev.tripdraw.presentation.member;
 
 import static dev.tripdraw.exception.auth.AuthExceptionType.AUTH_FAIL;
 
-import dev.tripdraw.application.MemberService;
+import dev.tripdraw.member.application.MemberService;
 import dev.tripdraw.dto.auth.LoginUser;
 import dev.tripdraw.exception.auth.AuthException;
 import jakarta.servlet.http.HttpServletRequest;
