+++ backend/src/test/java/dev/tripdraw/member/domain/MemberTest.java
@@ -1,8 +1,9 @@
-package dev.tripdraw.domain.member;
+package dev.tripdraw.member.domain;
 
 import static dev.tripdraw.domain.oauth.OauthType.KAKAO;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import dev.tripdraw.member.domain.Member;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
